Add float type casting

+++ SW1_HW1_IO_BASIC/20196487이준학_SW1_HW1_2.c
@@ -11,7 +11,7 @@ int main() {
 	if (!scanf("%c", &name)) exit(1);
 
 	int total_sc = kor_sc + eng_sc;
-	float average_sc = total_sc / 2;
+	float average_sc = (float)total_sc / 2;
 
 	printf("NAME ENGLISH KOREAN TOTAL AVERAGE\n");
 	printf("     SCORE   SCORE  SCORE SCORE\n");
